+++ Gateway/Http/Client/ResponseFactory.php
@@ -17,6 +17,8 @@
 
 namespace Mastercard\Mastercard\Gateway\Http\Client;
 
+use Laminas\Http\Response;
+
 class ResponseFactory
 {
     /**
@@ -27,10 +29,6 @@ class ResponseFactory
      */
     public function create($response)
     {
-        if (class_exists('\Zend_Http_Response')) {
-            return \Zend_Http_Response::fromString($response);
-        }else {
-            return \Laminas\Http\Response::fromString($response);
-        }
+        return Response::fromString($response);
     }
 }